+++ api/barrier_reports/constants.py
@@ -1,0 +1,79 @@
+BARRIER_FIELD_TO_REPORT_TITLE = {
+    "id": "id",
+    "code": "code",
+    "title": "Title",
+    "status": "Status",
+    "priority": "Old Priority",
+    "priority_level": "Priority",
+    "overseas_region": "Overseas Region",
+    "location": "Location",
+    "admin_areas": "Admin areas",
+    "sectors": "Sectors",
+    "product": "Product",
+    "term": "Term",
+    "categories": "Barrier categories",
+    "source": "Source",
+    "team_count": "Team count",
+    "economic_assessment_rating": "Economic assessment rating",
+    "economic_assessment_explanation": "Economic assessment explanation",
+    "value_to_economy": "Value to economy",
+    "import_market_size": "Import market size",
+    "valuation_assessment_rating": "Valuation assessment rating",
+    "valuation_assessment_midpoint": "Midpoint value",
+    "valuation_assessment_explanation": "Valuation assessment explanation",
+    "commercial_value": "Commercial value estimate",
+    "reported_on": "Reported Date",
+    "status_date": "Status effective from date",
+    "resolved_date": "Resolved Date",
+    "status_summary": "Status summary",
+    "modified_on": "Last updated",
+    "tags": "Tags",
+    "trade_direction": "Trade direction",
+    "estimated_resolution_date": "Estimated resolution date",
+    "proposed_estimated_resolution_date": "Proposed estimated resolution date",
+    "previous_estimated_resolution_date": "The previous estimate for resolution date",
+    "estimated_resolution_updated_date": "The last date the resolution date was re-estimated",
+    "estimated_resolution_date_change_reason": "The reason for change to the estimated resolution date",
+    "summary": "Summary",
+    "link": "Link",
+    "wto_has_been_notified": "WTO Notified",
+    "wto_should_be_notified": "WTO Should Notify",
+    "wto_committee_notified": "WTO Committee Notified",
+    "wto_committee_notification_link": "WTO Committee Notified Link",
+    "wto_member_states": "WTO Raised Members",
+    "wto_committee_raised_in": "WTO Raised Committee",
+    "wto_raised_date": "WTO Raised Date",
+    "wto_case_number": "WTO Case Number",
+    "commodity_codes": "HS commodity codes",
+    "first_published_on": "First published date",
+    "last_published_on": "Last published date",
+    "public_view_status": "Public view status",
+    "public_eligibility_summary": "Public eligibility summary",
+    # TODO: last_public_view_status_update takes too long to calculate
+    # on production, and needs to be denormalised. Will be addressed in MAR-940
+    "last_public_view_status_update": "Last public view status update",
+    "changed_since_published": "Changed since published",
+    "public_id": "Public ID",
+    "public_title": "Public title",
+    "public_summary": "Public summary",
+    "public_is_resolved": "Published as resolved",
+    "latest_publish_note": "Latest publish note",
+    "resolvability_assessment_time": "Resolvability assessment time",
+    "resolvability_assessment_effort": "Resolvability assessment effort",
+    "strategic_assessment_scale": "Strategic assessment scale",
+    "is_top_priority": "Is Top Priority",
+    "top_priority_status": "Top Priority Status",
+    "top_priority_summary": "Reason for Top Priority Status",
+    "is_resolved_top_priority": "Is Resolved Top Priority",
+    "government_organisations": "Related Organisations",
+    "progress_update_status": "Progress update status",
+    "progress_update_message": "Progress update message",
+    "progress_update_date": "Progress update date",
+    "progress_update_author": "Progress update author",
+    "progress_update_next_steps": "Progress update next steps",
+    "next_steps_items": "Progress update list of next steps",
+    "programme_fund_progress_update_milestones": "Programme fund milestones",
+    "programme_fund_progress_update_expenditure": "Programme fund expenditure",
+    "programme_fund_progress_update_date": "Programme fund date",
+    "programme_fund_progress_update_author": "Programme fund author",
+}
