+++ tests/barrier_reports/test_csv.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 
 import pytest
 
-from api.barriers.csv import _transform_csv_value
+from barrier_reports.csv import _transform_csv_value
 
 
 @pytest.mark.parametrize(
